TERMINADO

+++ SistemaActividades/Modelo/Actividad.cs
@@ -6,12 +6,16 @@ using System.Threading.Tasks;
 
 namespace SistemaActividades.Modelo
 {
+    /// <summary>
+    /// EVELYN PAMELA GUTIERREZ RUIZ 14/09/2026
+    /// ESTA CLASE REPRESENTA UNA ACTIVIDAD
+    /// </summary>
     public abstract class Actividad
     {
         // Propiedades
         public int Id { get; set; }
         public string Nombre { get; set; }
-        public double costo  { get; set; }
+        public decimal costo { get; set; }
        
 
         //propiedad de tipo Responsable
